[dnf-automatic] Wait for internet connection (RhBug:1816308)
The dnf-automatic* timers are set to 6AM by default. This works
fine for servers running 24x7 but there is great deal of probability
that workstations are suspended by that time. In that case the timers
are fired as soon as posible after the machine is resumed - even before
the network being properly set-up.
Now dnf-automatic waits until it can establish a connection to at
least one of enabled remote repositories.

https://bugzilla.redhat.com/show_bug.cgi?id=1816308